+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to the nUIget extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2026-02-10
+
+### Fixed
+
+- **Slow Installed Tab with Unreachable Sources**
+- **Redundant project.assets.json Parsing**
+- **Transitive Prefetch Network Contention**
+
 ## [1.1.0] - 2026-02-10
 
 ### Added